RDB archive: Fix Oracle time zone problem.
The sample.smpl_time is handled differently for MySQL, Postgres, Oracle.

In MySQL, it's stored as UTC. The `Timestamp` passed to
`PreparedStatement.setTimestamp(Timestamp)` is converted to UTC and
stored as UTC. On retrieval, `ResultSet.getTimestamp` converts back to
the local time zone with appropriate GMT offset.

In Postgres, we need to use a `TIMESTAMPTZ` datatype that stores the
stamp with timezone info.

Both MySQL and Postgres handle the fall transition from daylight savings
time back to standard time just fine.

In Oracle, even if we use `TIMESTAMP WITH TIMEZONE` for smpl_time,
Oracle JDBC will second-guess the Timestamp passed to
`setTimestamp(Timestamp)` and change the time offset. During the DST
change in the fall, time stamps will be written with the wrong GMT
offset.
The workaround is to use `setObject(OffsetDateTime)` because
OffsetDateTime is passed through and written as received.